Aging alignment with EMOD-Generic (#228)

+++ Eradication/IndividualEventContext.h
@@ -20,25 +20,23 @@ namespace Kernel
     {
         virtual const IIndividualHuman* GetIndividualHumanConst() const = 0;
 
-        // for example, individual attributes...
         virtual bool   IsPregnant()          const = 0;        
         
-        virtual double GetAge()              const = 0;
+        virtual float  GetAge()              const = 0;
         virtual int    GetGender()           const = 0;
-        virtual double GetMonteCarloWeight() const = 0;
+        virtual float  GetMonteCarloWeight() const = 0;
         virtual bool   IsPossibleMother()    const = 0;
         virtual bool   IsInfected()          const = 0;
         virtual bool   IsSymptomatic()       const = 0;
         virtual float  GetInfectiousness()   const = 0;
+
         virtual HumanStateChange GetStateChange(void) const = 0;
         virtual void   Die( HumanStateChange ) = 0; // for testing, quick results
         virtual suids::suid GetSuid() const = 0;
 
         // useful sub contexts
-        virtual IIndividualHumanInterventionsContext *GetInterventionsContext() const  = 0;
-        virtual INodeEventContext *GetNodeEventContext() = 0;
-        virtual IPKeyValueContainer * GetProperties() = 0;
-
-        // TODO: add more methods here to extend programmatic capability of event coordinators w.r.t. to individual state!
+        virtual IIndividualHumanInterventionsContext*  GetInterventionsContext() const  = 0;
+        virtual INodeEventContext*                     GetNodeEventContext()            = 0;
+        virtual IPKeyValueContainer*                   GetProperties()                  = 0;
     };
 }
